+++ src/rds_proxy_password_rotatation/adapter/aws_lambda_function.py
@@ -1,4 +1,15 @@
+from enum import Enum
 from pydantic import BaseModel, Field
+
+from aws_lambda_powertools.utilities.parser import event_parser
+from aws_lambda_powertools.utilities.typing import LambdaContext
+
+
+class RotationStep(Enum):
+    CREATE_SECRET = "create_secret"
+    SET_SECRET = "set_secret"
+    TEST_SECRET = "test_secret"
+    FINISH_SECRET = "finish_secret"
 
 
 class AwsSecretManagerRotationEvent(BaseModel):
@@ -11,7 +22,15 @@ class AwsSecretManagerRotationEvent(BaseModel):
 
     RotationToken – A unique identifier that indicates the source of the request. Required for secret rotation using an assumed role or cross-account rotation, in which you rotate a secret in one account by using a Lambda rotation function in another account. In both cases, the rotation function assumes an IAM role to call Secrets Manager and then Secrets Manager uses the rotation token to validate the IAM role identity.
     """
-    step: str = Field(alias='Step')
-    secretId: str = Field(alias='SecretId')
-    clientRequestToken: str = Field(alias='ClientRequestToken')
-    rotationToken: str = Field(alias='RotationToken')
+    step: RotationStep = Field(alias='Step')
+    secret_id: str = Field(alias='SecretId')
+    client_request_token: str = Field(alias='ClientRequestToken')
+    rotation_token: str = Field(alias='RotationToken')
+
+
+@event_parser(model=AwsSecretManagerRotationEvent)
+def lambda_handler(event: AwsSecretManagerRotationEvent, context: LambdaContext) -> None:
+    print(event)
+    print(context)
+
+    return event
